back/requirements.txt
django==6.0.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
numpy==2.2.5
psycopg2-binary==2.9.10
gunicorn==23.0.0
python-decouple==3.8
